Add likes styles
