+++ apps/erya/src/app/shared/components/modal/modal.component.html
@@ -1,0 +1,16 @@
+<div class="modal">
+  <div class="modal-background"></div>
+  <div class="modal-content">
+    <div class="columns is-centered">
+      <div class="column is-half">
+        <div class="box content">
+          <ng-content></ng-content>
+        </div>
+      </div>
+    </div>
+
+  </div>
+  <button class="modal-close is-large" aria-label="close"></button>
+</div>
+
+
