+++ apps/erya/src/assets/stylesheets/_main.scss
@@ -143,3 +143,7 @@ input.sense-input {
   opacity: 1;
   cursor: default;
 }
+
+.content figure.has-likes:not(:first-child) {
+  margin-top: 0;
+}
